Use explicit kubectl rather than alias for mk8s (#24)
* Use explicit kubectl rather than alias for mk8s

Workflows which expect kubectl to be runnable without the microk8s group
break after #23 due to the alias needing said group. Always using the
actual kubectl snap and copying the microk8s config over instead avoids
the issue.

* Fix microk8s.config requiring group as well

* Skip model / controller cleanup entirely for microk8s

It fails frequently and can take an inordinately long time even when it
does succeed, and we expect the entire environment to be cleaned up when
the runner finishes anyway.