Adds optional addons for microk8s
In order to be able to use the nginx-ingress-integrator charm and other
related charms, we need to enable ingress in microk8s.

This adds the microk8s-addons option, which will contain the addons to be
enabled in microk8s. If not given, the default addons will be added instead
(storage, dns, rbac).